Creating a GNSS function file
